alteracoes revisao time de apoio
