+++ pages/busca-ativa/gestantes/index.js
@@ -2,11 +2,8 @@ import {
   CardAlert,
   TituloTexto, 
   ButtonLight, 
-  PainelBuscaAtiva , 
-  ScoreCardGrid , 
-  Spinner, 
   ButtonPrint,
-  TabelaCitoImpressao,
+  TabelaHiperDiaImpressao,
   PanelSelector
 } from "@impulsogov/design-system";
 import React, { useState,useEffect } from 'react';
@@ -16,7 +13,10 @@ import { useRouter } from 'next/router';
 import { getData } from '../../../services/cms'
 import { LAYOUT } from '../../../utils/QUERYS'
 import { redirectHome } from "../../../helpers/redirectHome";
-import { colunasGestantes } from "../../../helpers/colunasGestantes";
+import { colunasGestantesEquipe } from "../../../helpers/colunasGestantes";
+import { colunasGestantesIndicadorUm } from "../../../helpers/colunasGestantesIndicadorUm";
+import { colunasGestantesIndicadorDois } from "../../../helpers/colunasGestantesIndicadorDois";
+import { colunasGestantesIndicadorTres } from "../../../helpers/colunasGestantesIndicadorTres";
 import { GraficoIndicadorUm, CardsGraficoIndicadorUm } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_1/grafico_indicador_1";
 import { tabelaGestantesEquipe , tabelaGestantesAPS } from "../../../services/busca_ativa/Gestantes";
 import { TabelaEquipeGestantesAtivas } from "../../../componentes/mounted/busca-ativa/gestantes/Equipe/tabelas/GestantesAtivas";
@@ -25,13 +25,10 @@ import { TabelaEquipeGestantesSemDUM } from "../../../componentes/mounted/busca-
 import { CardsEquipe } from "../../../componentes/mounted/busca-ativa/gestantes/Equipe/cardsEquipe";
 import { IndicadorUmTabelaGestantesAtivas } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_1/tabelas/GestantesAtivas";
 import { IndicadorUmTabelaGestantesEncerradas } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_1/tabelas/GestantesEncerradas";
-import { IndicadorUmTabelaGestantesSemDUM } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_1/tabelas/GestantesSemDum";
 import { IndicadorDoisTabelaGestantesAtivas } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_2/tabelas/GestantesAtivas";
 import { IndicadorDoisTabelaGestantesEncerradas } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_2/tabelas/GestantesEncerradas";
-import { IndicadorDoisTabelaGestantesSemDUM } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_2/tabelas/GestantesSemDum";
 import { IndicadorTresTabelaGestantesAtivas } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_3/tabelas/GestantesAtivas";
 import { IndicadorTresTabelaGestantesEncerradas } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_3/tabelas/GestantesEncerradas";
-import { IndicadorTresTabelaGestantesSemDUM } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_3/tabelas/GestantesSemDum";
 import { IndicadorUmCardsGestantesAtivas } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_1/cards/cardsGestantesAtivas";
 import { IndicadorUmCardsGestantesEncerradas } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_1/cards/cardsGestantesEncerradas";
 import { IndicadorDoisCardsGestantesAtivas } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_2/cards/cardsGestantesAtivas";
@@ -40,8 +37,6 @@ import { IndicadorTresCardsGestantesAtivas } from "../../../componentes/mounted/
 import { IndicadorTresCardsGestantesEncerradas } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_3/cards/cardsGestantesEncerradas";
 import { TabelaGestantesSemDUM } from "../../../componentes/mounted/busca-ativa/gestantes/APS/GestantesSemDUM/GestantesSemDum";
 
-import status_usuario_descricao  from "../../../data/StatusAcompanhamento.json" assert { type: 'json' };
-import identificacao_exame_hiv_sifilis  from "../../../data/identificacao_exame_hiv_sifilis.json" assert { type: 'json' };
 import { CardsGraficoIndicadorDois, GraficoIndicadorDois } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_2/grafico_indicador_2";
 import { CardsGraficoIndicadorTres, GraficoIndicadorTres } from "../../../componentes/mounted/busca-ativa/gestantes/APS/indicador_3/grafico_indicador_3";
 import { CardsAPS } from "../../../componentes/mounted/busca-ativa/gestantes/APS/cardsAPS";
@@ -93,253 +88,14 @@ useEffect(()=>{
     setTabelaDataEquipe(response.data)
 })},[session]) 
 const [tabelaData, setTabelaData] = useState([]);
-const datefiltrosGestantes = [
-    "gestacao_data_dpp",
-    "consulta_prenatal_ultima_data",
-]
-const rotulosfiltrosGestantes = [
-    "NOMES DE A-Z",
-    "NOME DO PROFISSIONAL RESPONSÁVEL DE A-Z",
-    "DPP MAIS PRÓXIMA",
-    "IG ATUAL MAIOR",
-    "MAIOR Nº TOTAL DE CONSULTAS",
-    "DATA MAIS RECENTE DA ÚLTIMA CONSULTA"
-]
-const IDFiltrosGestantes = {
-    "NOMES DE A-Z" : "cidadao_nome",
-    "NOME DO PROFISSIONAL RESPONSÁVEL DE A-Z" : "acs_nome",
-    "DPP MAIS PRÓXIMA" : "gestacao_data_dpp",
-    "IG ATUAL MAIOR" : "gestacao_idade_gestacional_atual",
-    "MAIOR Nº TOTAL DE CONSULTAS" : "consultas_pre_natal_validas",
-    "DATA MAIS RECENTE DA ÚLTIMA CONSULTA" : "consulta_prenatal_ultima_data"
-}   
-const IDFiltrosOrdenacaoGestantes = {
-    "NOMES DE A-Z" : "asc",
-    "NOME DO PROFISSIONAL RESPONSÁVEL DE A-Z" : "asc",
-    "DPP MAIS PRÓXIMA" : "asc",
-    "IG ATUAL MAIOR" : "desc",
-    "MAIOR Nº TOTAL DE CONSULTAS" : "desc",
-    "DATA MAIS RECENTE DA ÚLTIMA CONSULTA" : "desc"
+const colunasImpressao = {
+  0 : colunasGestantesIndicadorUm,
+  1 : colunasGestantesIndicadorDois,
+  3 : colunasGestantesIndicadorTres,
+  4 : colunasGestantesIndicadorUm
 }
 if(session){  
   if(session.user.perfis.includes(9)){
-      const CardsChild = tabelaDataEquipe ? <ScoreCardGrid
-      valores={[
-          {
-              descricao: 'Total de gestantes',
-              valor: [...new Set(tabelaDataEquipe.map(item => item.chave_id_gestacao))].length
-          },
-          {
-              descricao: 'Total de gestantes ativas',
-              valor: tabelaDataEquipe.filter(item=> item.id_status_usuario == 8).length
-          },
-          {
-            descricao: 'Total de getantes ativas contabilizando para os três indicadores',
-            valor: tabelaDataEquipe.filter(item=>{ 
-            return(
-              item.id_status_usuario == 8
-              && item.consultas_pre_natal_validas >= 6
-              && item.gestacao_idade_gestacional_primeiro_atendimento <= 12
-              && item.id_exame_hiv_sifilis == 4
-              && item.id_atendimento_odontologico == 1
-            )}).length
-          },
-          {
-            descricao: 'Gestantes ativas, com 6 ou mais consultas, com a primeira consulta realizada até a 12ª semana de gestação',
-            valor: tabelaDataEquipe.filter((item)=>{ 
-            return (
-              item.id_status_usuario == 8
-              && item.consultas_pre_natal_validas >=6
-              && item.gestacao_idade_gestacional_primeiro_atendimento <= 12
-            )}).length
-          },
-          {
-            descricao: 'Gestantes ativas, com atendimento odontológico realizado',
-            valor: tabelaDataEquipe.filter((item)=>{ 
-            return (
-              item.id_status_usuario == 8
-              && item.id_atendimento_odontologico == 1) 
-            }).length
-          },
-          {
-            descricao: 'Gestantes ativas, com os dois exames realizados e identificados',
-            valor: tabelaDataEquipe.filter((item)=>{ 
-            return ( 
-              item.id_status_usuario == 8
-              && item.id_exame_hiv_sifilis == 4)
-            }).length
-          },
-    ]}
-      /> : <Spinner/>
-  const tabelaDataEquipeGestantesAtivas = tabelaDataEquipe.filter(item=>item.id_status_usuario == 8)
-  const TabelaChildGestantesAtivas = tabelaDataEquipeGestantesAtivas && tabelaDataEquipeGestantesAtivas.length>0 && tabelaDataEquipe && tabelaData ? 
-  <>
-  <PainelBuscaAtiva
-      key="TabelaChildGestantesAtivas"
-      dadosFiltros={[
-        {
-          data: [...new Set(tabelaDataEquipeGestantesAtivas.map(item => item.acs_nome))],
-          filtro: 'acs_nome',
-          rotulo: 'Filtrar por Profissional Responsável'
-        },
-        {
-            data: [...new Set(tabelaDataEquipeGestantesAtivas.map(item => item.consultas_pre_natal_validas.toString()))],
-            filtro: 'consultas_pre_natal_validas',
-            rotulo: 'Filtrar por numero de consultas'
-        },
-        {
-            data: [...new Set(tabelaDataEquipeGestantesAtivas.map(item => item.id_atendimento_odontologico.toString()))],
-            labels : {1 : "Sim", 2 : "Não"},
-            filtro: 'id_atendimento_odontologico',
-            rotulo: 'Filtrar por atendimento odontológico'
-        },
-        {
-            data: [...new Set(tabelaDataEquipeGestantesAtivas.map(item => item.id_exame_hiv_sifilis.toString()))],
-            labels : [...new Set(identificacao_exame_hiv_sifilis.identificacao_exame_hiv_sifilis.map(item=> item.exame_hiv_sifilis_descricao))],
-            filtro: 'id_exame_hiv_sifilis',
-            rotulo: 'Filtrar por identificação do exame de sífilis e HIV'
-        },
-        {
-          data: [...new Set(tabelaDataEquipeGestantesAtivas.map(item => item.gestacao_quadrimestre))],
-          filtro: 'gestacao_quadrimestre',
-          rotulo: 'Filtrar por quadrimestre'
-        },
-      ]}
-      painel="gestantes"
-      tabela={{
-      colunas: colunasGestantes,
-      data:tabelaDataEquipeGestantesAtivas
-      }}
-      data={tabelaData}
-      setData={setTabelaData}
-      datefiltros={datefiltrosGestantes}
-      IDFiltros={IDFiltrosGestantes}
-      rotulosfiltros={rotulosfiltrosGestantes}    
-      IDFiltrosOrdenacao={IDFiltrosOrdenacaoGestantes}
-      rowHeight={65}
-      atualizacao = {new Date(tabelaDataEquipeGestantesAtivas.reduce((maisRecente, objeto) => {
-        const dataAtual = new Date(objeto.dt_registro_producao_mais_recente);
-        const dataMaisRecenteAnterior = new Date(maisRecente);
-        return dataAtual > dataMaisRecenteAnterior ? objeto.dt_registro_producao_mais_recente : maisRecente
-        }, "2000-01-01")).toLocaleString('pt-BR', { 
-        timeZone: 'UTC',
-        year: 'numeric',
-        month: '2-digit',
-        day: '2-digit'
-        })}
-  /></> : <Spinner />
-  
-  const tabelaDataEquipeGestantesSemDUM = tabelaDataEquipe.filter(item=>item.id_status_usuario == 11)
-  const TabelaChildGestantesSemDUM = tabelaDataEquipeGestantesSemDUM && tabelaDataEquipeGestantesSemDUM.length>0 && tabelaDataEquipe && tabelaData ? 
-  <>
-  <PainelBuscaAtiva
-      key="TabelaChildGestantesSemDUM"
-      dadosFiltros={[
-        {
-          data: [...new Set(tabelaDataEquipeGestantesSemDUM.map(item => item.acs_nome))],
-          filtro: 'acs_nome',
-          rotulo: 'Filtrar por Profissional Responsável'
-        },
-        {
-            data: [...new Set(tabelaDataEquipeGestantesSemDUM.map(item => item.consultas_pre_natal_validas?.toString()))],
-            filtro: 'consultas_pre_natal_validas',
-            rotulo: 'Filtrar por numero de consultas'
-        },
-        {
-            data: [...new Set(tabelaDataEquipeGestantesSemDUM.map(item => item.id_atendimento_odontologico?.toString()))],
-            filtro: 'id_atendimento_odontologico',
-            rotulo: 'Filtrar por atendimento odontológico'
-        },
-        {
-            data: [...new Set(tabelaDataEquipeGestantesSemDUM.map(item => item.id_exame_hiv_sifilis?.toString()))],
-            filtro: 'id_exame_hiv_sifilis',
-            rotulo: 'Filtrar por identificação do exame de sífilis e HIV'
-        },
-        {
-          data: [...new Set(tabelaDataEquipeGestantesSemDUM.map(item => item.gestacao_quadrimestre))],
-          filtro: 'gestacao_quadrimestre',
-          rotulo: 'Filtrar por quadrimestre'
-        },
-      ]}
-      painel="gestantes"
-      tabela={{
-      colunas: colunasGestantes,
-      data:tabelaDataEquipeGestantesSemDUM
-      }}
-      data={tabelaData}
-      setData={setTabelaData}
-      datefiltros={datefiltrosGestantes}
-      IDFiltros={IDFiltrosGestantes}
-      rotulosfiltros={rotulosfiltrosGestantes}    
-      IDFiltrosOrdenacao={IDFiltrosOrdenacaoGestantes}
-      atualizacao = {new Date(tabelaDataEquipeGestantesSemDUM.reduce((maisRecente, objeto) => {
-        const dataAtual = new Date(objeto.dt_registro_producao_mais_recente);
-        const dataMaisRecenteAnterior = new Date(maisRecente);
-        return dataAtual > dataMaisRecenteAnterior ? objeto.dt_registro_producao_mais_recente : maisRecente
-        }, "2000-01-01")).toLocaleString('pt-BR', { 
-        timeZone: 'UTC',
-        year: 'numeric',
-        month: '2-digit',
-        day: '2-digit'
-        })}
-
-  /></> : <Spinner/>
-
-  const tabelaDataEquipeGestantesEncerradas = tabelaDataEquipe?.filter(item=>item.id_status_usuario == 9)
-  const TabelaChildGestantesEncerradas = tabelaDataEquipeGestantesEncerradas && tabelaDataEquipeGestantesEncerradas.length>0 && tabelaDataEquipe && tabelaData ? 
-  <>
-  <PainelBuscaAtiva
-      key="TabelaChildGestantesEncerradas"
-      dadosFiltros={[
-        {
-          data: [...new Set(tabelaDataEquipeGestantesEncerradas.map(item => item.acs_nome))],
-          filtro: 'acs_nome',
-          rotulo: 'Filtrar por Profissional Responsável'
-        },
-        {
-            data: [...new Set(tabelaDataEquipeGestantesEncerradas.map(item => item.consultas_pre_natal_validas.toString()))],
-            filtro: 'consultas_pre_natal_validas',
-            rotulo: 'Filtrar por numero de consultas'
-        },
-        {
-            data: [...new Set(tabelaDataEquipeGestantesEncerradas.map(item => item.id_atendimento_odontologico.toString()))],
-            filtro: 'id_atendimento_odontologico',
-            rotulo: 'Filtrar por atendimento odontológico'
-        },
-        {
-            data: [...new Set(tabelaDataEquipeGestantesEncerradas.map(item => item.id_exame_hiv_sifilis.toString()))],
-            filtro: 'id_exame_hiv_sifilis',
-            rotulo: 'Filtrar por identificação do exame de sífilis e HIV'
-        },
-        {
-          data: [...new Set(tabelaDataEquipeGestantesEncerradas.map(item => item.gestacao_quadrimestre))],
-          filtro: 'gestacao_quadrimestre',
-          rotulo: 'Filtrar por quadrimestre'
-        },
-      ]}
-      painel="gestantes"
-      tabela={{
-      colunas: colunasGestantes,
-      data:tabelaDataEquipeGestantesEncerradas
-      }}
-      data={tabelaData}
-      setData={setTabelaData}
-      datefiltros={datefiltrosGestantes}
-      IDFiltros={IDFiltrosGestantes}
-      rotulosfiltros={rotulosfiltrosGestantes}    
-      IDFiltrosOrdenacao={IDFiltrosOrdenacaoGestantes}
-      atualizacao = {new Date(tabelaDataEquipeGestantesEncerradas.reduce((maisRecente, objeto) => {
-        const dataAtual = new Date(objeto.dt_registro_producao_mais_recente);
-        const dataMaisRecenteAnterior = new Date(maisRecente);
-        return dataAtual > dataMaisRecenteAnterior ? objeto.dt_registro_producao_mais_recente : maisRecente
-        }, "2000-01-01")).toLocaleString('pt-BR', { 
-        timeZone: 'UTC',
-        year: 'numeric',
-        month: '2-digit',
-        day: '2-digit'
-        })}
-
-  /></> : <Spinner/>
   const Children = [[
     [
         <CardsEquipe tabelaDataEquipe={tabelaDataEquipe}/>,
@@ -367,12 +123,12 @@ if(session){
           url: 'https://media.graphassets.com/8NbkQQkyRSiouNfFpLOG'}} 
           label="VOLTAR" link="/inicio"/>
       {
-          tabelaDataEquipe &&
+          tabelaData &&
           <div style={{marginLeft:"auto"}}>
           <ButtonPrint
               label="CLIQUE AQUI PARA IMPRIMIR"
               escala="0.78"
-              child={<TabelaCitoImpressao data={tabelaData} colunas={colunasGestantes} status_usuario_descricao={status_usuario_descricao}/>}
+              child={<TabelaHiperDiaImpressao data={tabelaData} colunas={colunasGestantesEquipe}/>}
           />
           </div>
       }
@@ -530,7 +286,7 @@ if(session){
             <ButtonPrint
                 label="CLIQUE AQUI PARA IMPRIMIR"
                 escala="0.78"
-                child={<TabelaCitoImpressao data={tabelaData} colunas={colunasGestantes} status_usuario_descricao={status_usuario_descricao}/>}
+                child={<TabelaHiperDiaImpressao data={tabelaData} colunas={colunasImpressao[activeTitleTabIndex]}/>}
             />
             </div>
         }
@@ -604,8 +360,9 @@ if(session){
                   [
                     {
                       label: 'GESTANTES SEM DUM'
-                    },
-                  ],
+                    },    
+                  ]
+
                   ]}
               titles={[
                 {
